Generate adaptors and proxies with virtual destructor
The generated classes, when inherited by a base class a la `class MyObject : public sdbus::AdaptorInterfaces<my::ns::MyAdaptor>`, the generated adaptor/proxy does not have a virtual destructor. This causes issues when storing the created MyObject as a MyAdaptor, where upon destruction the destructor of MyObject is never called.