Add finished flag to CallData.
The previous fix consisted of adding CallData to the pendingAsyncCalls
before calling message.send, and then removing it if message.send fails.
Instead, add a finished flag to CallData that is set to true whenever
removeCall is called. In that case we don't need to add it to the list in
addCall.
The resulting code in calMethod is more straight forward.

Signed-off-by: Anthony Brandon <anthony@amarulasolutions.com>